tarea: reconocimiento impresora

+++ src/main.cpp
@@ -1,12 +1,7 @@
-﻿#include <cpprest/http_listener.h>
-#include <cpprest/json.h>
-#include <cpprest/uri.h>
-#include <cpprest/asyncrt_utils.h>
-#include <cpprest/filestream.h>
-#include <iostream>
-#include <vector>
-#include <string>
-#include <filesystem>
+﻿#include "pch.h"
+#include "main.h"
+#include "logger.h"
+#include "check_printer.h"
 
 using namespace web;
 using namespace web::http;
@@ -35,20 +30,25 @@ void handle_print_request(http_request request) {
 			out_file.write(reinterpret_cast<const char*>(byte_array.data()), byte_array.size());
 			out_file.close();
 
-			std::cout << "Imagen guardada como " << nombre_archivo << std::endl;
-			// Imprime imagen 
-			int status_code = system(comando.c_str());
-			if (status_code == 0) {
-				//Elimina la imagen
-				if (std::filesystem::remove(nombre_archivo))
-					std::cout << "Imagen " << nombre_archivo << " Eliminada.\n";
-				else
-					std::cout << "Imagen " << nombre_archivo << " no eliminada " << std::endl;
-				request.reply(status_codes::OK, U("Imagen correctamente recibida e impresa."));
+			spdlog::info("Imagen guardada como {}", nombre_archivo);
+			if (is_zebra_connected()) {
+				// Imprime imagen 
+				int status_code = system(comando.c_str());
+				if (status_code == 0) {
+					//Elimina la imagen
+					if (std::filesystem::remove(nombre_archivo))
+						spdlog::warn("Imagen {} eliminada.", nombre_archivo);
+					else
+						spdlog::warn("Imagen {} no eliminada ", nombre_archivo);
+					request.reply(status_codes::OK, U("Imagen correctamente recibida e impresa."));
+				}
+				else {
+					request.reply(status_codes::OK, U("Imagen correctamente recibida, guardada pero tiene que imprimir manualmente."));
+				}
 			}
-			else {
-				request.reply(status_codes::OK, U("Imagen correctamente recibida, guardada pero tiene que imprimir manualmente."));
-			}
+			else
+				request.reply(status_codes::ExpectationFailed, U("Impresora Zebra no conectada."));
+
 		}
 		catch (const std::exception& e) {
 			ucout << U("Error decodificando la cadena caracteres: ") << e.what() << std::endl;
@@ -65,6 +65,7 @@ void handle_print_request(http_request request) {
 }
 
 int main() {
+	rotate_spdlog();
 	uri_builder uri(U("http://*:3000")); // Listen on all available network interfaces
 	auto addr = uri.to_uri().to_string();
 	http_listener listener(addr);
@@ -73,7 +74,7 @@ int main() {
 	try {
 		listener
 			.open()
-			.then([&listener]() { std::wcout << L"Iniciando servidor en la dirección: " << listener.uri().to_string() << std::endl; })
+			.then([&listener]() { spdlog::info("Iniciando servidor en la dirección: {}", utility::conversions::to_utf8string(listener.uri().to_string())); })
 			.wait();
 
 		std::string line;
@@ -83,5 +84,6 @@ int main() {
 		std::cerr << "Error: " << e.what() << std::endl;
 	}
 
+	spdlog::info("Servidor finalizado");
 	return 0;
 }
